Replaced all occurences of the datatype 'long' with 'int', leaving types 'long long' as they were. This is guaranteed to work, because the 'long' datatype was 32 bit on a 32 bit architecture. As 'int' it will remain 32 bit even on a 64 bit machine.
Almost certainly the data aquisition does not work now, because in the
original code pointers were used for arithmetic which now broke with the
64 bit architecture. Instead of using pointers the datatype 'int' is used.
The change compiles but the functioning of the code is not guaranteed.

Removed the 'WinTypes.h' header that originates from an old version of the
FTDI2XX driver and conflicted with the library.

With these changes the program 'psi46expert' seems to work, tests were made
with 'PixelAlive', 'PreTest', and 'SCurve' which all succeeded.


git-svn-id: https://svn.cern.ch/reps/psi46expert/branches/64bit@6 4525493e-7705-40b1-a816-d608a930855b

+++ src/interface/Delay.h
@@ -13,14 +13,14 @@ class Delay {
 
 // == Delays ================================================
 	void Mdelay(unsigned int ms);
-	void Udelay(unsigned long us);
+	void Udelay(unsigned int us);
 
 // == Calibration ===========================================
 	int Median(unsigned int n, int x[]);
 	int Cpu_MHz();
 	int Calibrate();
 	int GetCPU_MHz();
-	unsigned long Usecs();
+	unsigned int Usecs();
 	QWORD Rdtsc();
 
 // == Runtime analysis =======================================
@@ -30,7 +30,7 @@ class Delay {
 	
  protected:
 
- 	 unsigned long uscount, usecsStart;
+ 	 unsigned int uscount, usecsStart;
  
 };
